Turn ChainInformation::from_chain_spec into ChainSpec::as_chain_information
