+++ src/chain_spec.rs
@@ -34,11 +34,15 @@
 //! - Multiple other miscellaneous information.
 //!
 
-use crate::chain::chain_information::{
-    BabeEpochInformation, ChainInformation, ChainInformationConsensus, ChainInformationFinality,
-    ValidChainInformation,
+use crate::{
+    chain::chain_information::{
+        aura_config, babe_config, BabeEpochInformation, ChainInformation,
+        ChainInformationConsensus, ChainInformationFinality, ValidChainInformation,
+    },
+    finality::grandpa,
 };
-use alloc::{string::String, vec::Vec};
+
+use alloc::{borrow::ToOwned as _, string::String, vec::Vec};
 use core::{convert::TryInto as _, num::NonZeroU64};
 
 mod light_sync_state;
@@ -151,6 +155,85 @@ impl ChainSpec {
         Ok(ChainSpec { client_spec })
     }
 
+    /// Builds the [`ChainInformation`] corresponding to the genesis block contained in this chain
+    /// spec.
+    pub fn as_chain_information(&self) -> Result<ChainInformation, FromGenesisStorageError> {
+        let consensus = {
+            let aura_genesis_config =
+                aura_config::AuraGenesisConfiguration::from_genesis_storage(|k| {
+                    self.genesis_storage_value(k).map(|v| v.to_owned())
+                });
+
+            let babe_genesis_config =
+                babe_config::BabeGenesisConfiguration::from_genesis_storage(|k| {
+                    self.genesis_storage_value(k).map(|v| v.to_owned())
+                });
+
+            match (aura_genesis_config, babe_genesis_config) {
+                (Ok(aura_genesis_config), Err(err)) if err.is_function_not_found() => {
+                    ChainInformationConsensus::Aura {
+                        finalized_authorities_list: aura_genesis_config.authorities_list,
+                        slot_duration: aura_genesis_config.slot_duration,
+                    }
+                }
+                (Err(err), Ok(babe_genesis_config)) if err.is_function_not_found() => {
+                    ChainInformationConsensus::Babe {
+                        slots_per_epoch: babe_genesis_config.slots_per_epoch,
+                        finalized_block_epoch_information: None,
+                        finalized_next_epoch_transition: BabeEpochInformation {
+                            epoch_index: 0,
+                            start_slot_number: None,
+                            authorities: babe_genesis_config.epoch0_information.authorities,
+                            randomness: babe_genesis_config.epoch0_information.randomness,
+                            c: babe_genesis_config.epoch0_configuration.c,
+                            allowed_slots: babe_genesis_config.epoch0_configuration.allowed_slots,
+                        },
+                    }
+                }
+                (Err(err1), Err(err2))
+                    if err1.is_function_not_found() && err2.is_function_not_found() =>
+                {
+                    // TODO: seems a bit risky to automatically fall back to this?
+                    ChainInformationConsensus::AllAuthorized
+                }
+                (Err(error), _) => {
+                    // Note that Babe might have produced an error as well, which is intentionally
+                    // ignored here in order to not make the API too complicated.
+                    return Err(FromGenesisStorageError::AuraConfigLoad(error));
+                }
+                (_, Err(error)) => {
+                    return Err(FromGenesisStorageError::BabeConfigLoad(error));
+                }
+                (Ok(_), Ok(_)) => {
+                    return Err(FromGenesisStorageError::MultipleConsensusAlgorithms);
+                }
+            }
+        };
+
+        let finality = {
+            let grandpa_genesis_config =
+                grandpa::chain_config::GrandpaGenesisConfiguration::from_genesis_storage(|k| {
+                    self.genesis_storage_value(k).map(|v| v.to_owned())
+                });
+
+            match grandpa_genesis_config {
+                Ok(grandpa_genesis_config) => ChainInformationFinality::Grandpa {
+                    after_finalized_block_authorities_set_id: 0,
+                    finalized_scheduled_change: None,
+                    finalized_triggered_authorities: grandpa_genesis_config.initial_authorities,
+                },
+                Err(error) if error.is_function_not_found() => ChainInformationFinality::Outsourced,
+                Err(error) => return Err(FromGenesisStorageError::GrandpaConfigLoad(error)),
+            }
+        };
+
+        Ok(ChainInformation {
+            finalized_block_header: crate::calculate_genesis_block_header(self),
+            consensus,
+            finality,
+        })
+    }
+
     /// Returns the name of the chain. Meant to be displayed to the user.
     pub fn name(&self) -> &str {
         &self.client_spec.name
@@ -248,6 +331,19 @@ impl ChainSpec {
 /// Error that can happen when parsing a chain spec JSON.
 #[derive(Debug, derive_more::Display)]
 pub struct ParseError(serde_json::Error);
+
+/// Error when building the chain information from the genesis storage.
+#[derive(Debug, derive_more::Display)]
+pub enum FromGenesisStorageError {
+    /// Error when retrieving the GrandPa configuration.
+    GrandpaConfigLoad(grandpa::chain_config::FromGenesisStorageError),
+    /// Error when retrieving the Aura algorithm configuration.
+    AuraConfigLoad(aura_config::FromGenesisStorageError),
+    /// Error when retrieving the Babe algorithm configuration.
+    BabeConfigLoad(babe_config::FromGenesisStorageError),
+    /// Multiple consensus algorithms have been detected.
+    MultipleConsensusAlgorithms,
+}
 
 #[cfg(test)]
 mod tests {
